requirements.txt
requests
pandas
sqlalchemy
python-dotenv
pyyaml
schedule
pytest